+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - changed: USP/Legacy Landing probability set to 50/50
 - changed: Use memos for Thorchain Savers withdrawals
 - changed: Allow wallet creation in WalletConnect wallet picker
+- changed: Improve transition to stake overview scene with preloaded data
+- changed: Remove warning dropdown if stake positions fail to load
 - fixed: Various margin styling alignments
 - fixed: Long delay updating exchange rates after login
 - fixed: Unable to claim unstaked TRX funds in some circumstances